pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "numpy>=1.21.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bhlff"
version = "0.1.0"
description = "7D Phase Field Theory Implementation for Elementary Particles"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Vasiliy Zdanovskiy", email = "vasilyvz@gmail.com"}
]
maintainers = [
    {name = "Vasiliy Zdanovskiy", email = "vasilyvz@gmail.com"}
]
keywords = [
    "phase field theory",
    "fractional calculus",
    "elementary particles",
    "7d physics",
    "topological defects",
    "solitons",
    "computational physics"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "h5py>=3.1.0",
    "pandas>=1.3.0",
    "tqdm>=4.62.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=12.0.0",
    "pytest>=6.2.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.4.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
    "pre-commit>=2.17.0",
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.17.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]
docs = [
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.17.0",
    "sphinx-autodoc-typehints>=1.12.0",
    "sphinx-copybutton>=0.5.0",
]
visualization = [
    "plotly>=5.0.0",
    "mayavi>=4.7.0",
    "vtk>=9.1.0",
    "pyvista>=0.35.0",
]
performance = [
    "numba>=0.56.0",
    "cupy-cuda11x>=10.0.0; platform_machine=='x86_64'",
    "cupy-cuda12x>=12.0.0; platform_machine=='x86_64'",
]

[project.urls]
Homepage = "https://github.com/vasilyvz/bhlff"
Documentation = "https://bhlff.readthedocs.io"
Repository = "https://github.com/vasilyvz/bhlff.git"
Issues = "https://github.com/vasilyvz/bhlff/issues"
Changelog = "https://github.com/vasilyvz/bhlff/blob/main/CHANGELOG.md"

[project.scripts]
bhlff = "bhlff.cli.main:main"
bhlff-run = "bhlff.cli.run:main"
bhlff-analyze = "bhlff.cli.analyze:main"
bhlff-report = "bhlff.cli.report:main"
bhlff-cuda-smoke = "bhlff.cli.cuda_smoke:main"
bhlff-quench-detect = "bhlff.cli.quench_detect:main"

[tool.setuptools.packages.find]
include = ["bhlff*"]

[tool.setuptools.package-data]
bhlff = [
    "configs/*.json",
    "configs/**/*.json",
    "data/*.h5",
    "data/*.npz",
    "templates/*.html",
    "templates/*.tex",
]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["bhlff"]
known_third_party = ["numpy", "scipy", "matplotlib", "h5py", "pandas"]

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "numba.*",
    "cupy.*",
    "mayavi.*",
    "vtk.*",
    "pyvista.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "-v",
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "level_a: marks tests for level A functionality",
    "level_b: marks tests for level B functionality",
    "level_c: marks tests for level C functionality",
    "level_d: marks tests for level D functionality",
    "level_e: marks tests for level E functionality",
    "level_f: marks tests for level F functionality",
    "level_g: marks tests for level G functionality",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/bhlff"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Flake8 configuration
[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg-info",
]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:S101",
]
